Extract GTop loading methods
